server/requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf_spectacular==0.26.0
psycopg2-binary>=2.9.1,<2.9.2
django-jazzmin==2.6.0
gunicorn==20.1.0
whitenoise==6.4.0
black==23.1.0
flake8==6.0.0
djangorestframework-simplejwt==5.2.2
wikipedia-api==0.5.8
pysbd==0.3.4